+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.40.0] - 2023-10-19
 ### Added
 - It is now possible to filter papers by `visibility`, `research_field` and optionally `include_subfields`
   (See: [!668](https://gitlab.com/TIBHannover/orkg/orkg-backend/-/merge_requests/668))
@@ -524,7 +526,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - This CHANGELOG file. Finally!
 
-[unreleased]: https://gitlab.com/TIBHannover/orkg/orkg-backend/-/compare/0.39.0...master
+[unreleased]: https://gitlab.com/TIBHannover/orkg/orkg-backend/-/compare/0.40.0...master
+[0.40.0]: https://gitlab.com/TIBHannover/orkg/orkg-backend/-/tags/0.40.0
 [0.39.0]: https://gitlab.com/TIBHannover/orkg/orkg-backend/-/tags/0.39.0
 [0.38.0]: https://gitlab.com/TIBHannover/orkg/orkg-backend/-/tags/0.38.0
 [0.37.0]: https://gitlab.com/TIBHannover/orkg/orkg-backend/-/tags/0.37.0